refactor: create new users module
