+++ users/developer_level_permission.go
@@ -1,4 +1,4 @@
-package googleplay
+package users
 
 type DeveloperLevelPermission string
