+++ users/user.go
@@ -1,4 +1,4 @@
-package googleplay
+package users
 
 type User struct {
 	Name                       string                     `json:"name"`
